+++ breeze-data/src/test/kotlin/com/windea/breezeframework/data/extensions/CoreExtensionsKtTest.kt
@@ -9,29 +9,20 @@ class CoreExtensionsKtTest {
 	@Test
 	fun testEquals() {
 		val foo1 = Foo(123, "abc")
-		val foo2 = Foo(233, "abc")
-		val foo3 = Foo(123, "abc")
-		assertNotEquals(foo1, foo2)
-		assertEquals(foo1, foo3)
+		val foo2 = Foo(123, "abc")
+		val foo3 = Foo(233, "abc")
+		assertTrue(equalsByCondition(foo1, foo2) { a, b -> a.a == b.a && a.b == b.b })
+		assertFalse(equalsByCondition(foo1, foo3) { a, b -> a.a == b.a && a.b == b.b })
+		assertTrue(equalsBySelect(foo1, foo2) { arrayOf(a, b) })
+		assertFalse(equalsBySelect(foo1, foo3) { arrayOf(a, b) })
 	}
 	
 	@Test
 	fun testToString() {
 		val foo = Foo(123, "abc")
-		assertEquals(foo.toString(), "Foo(a=123, b='abc')")
+		assertEquals(toStringBySelect(foo) { arrayOf("a" to a, "b" to b) }, "Foo(a=123, b='abc')")
+		assertEquals(toStringBySelectRef(foo) { arrayOf(::a, ::b) }, "Foo(a=123, b='abc')")
 	}
 }
 
-class Foo(val a: Int, val b: String) {
-	override fun equals(other: Any?): Boolean {
-		return equalsByPredicate(this, other) { a, b -> a.a == b.a && a.b == b.b }
-	}
-	
-	override fun hashCode(): Int {
-		return hashcodeBySelect(this) { arrayOf(it.a, it.b) }
-	}
-	
-	override fun toString(): String {
-		return toStringBySelect(this) { arrayOf(it::a to it.a, it::b to it.b) }
-	}
-}
+class Foo(val a: Int, val b: String)